@@ -6,31 +6,38 @@
   },
   "private": true,
   "devDependencies": {
-    "react-test-renderer": "^17.0.2",
-    "redux-mock-store": "^1.5.4",
+    "@wojtekmaj/enzyme-adapter-react-17": "^0.6.7",
+    "core-js": "^3.25.1",
+    "enzyme": "^3.11.0",
     "eslint-config-prettier": "^8.3.0",
     "eslint-plugin-react": "^7.26.1",
-    "prettier": "2.4.1"
+    "prettier": "2.4.1",
+    "react-test-renderer": "^17.0.2",
+    "redux-mock-store": "^1.5.4"
   },
   "dependencies": {
+    "@babel/core": "^7.18.13",
+    "@babel/plugin-proposal-private-property-in-object": "^7.18.6",
+    "@babel/plugin-syntax-flow": "^7.18.6",
+    "@babel/plugin-transform-react-jsx": "^7.18.10",
     "bootstrap": "^4.6.0",
     "bootstrap-icons": "^1.7.1",
     "history": "^4.6.3",
+    "jquery": "^3.6.1",
     "marked": "^0.3.6",
-    "sass": "^1.45.0",
+    "postcss": "^8.4.16",
     "prop-types": "^15.5.10",
-    "react": "^16.3.0",
-    "react-dom": "^16.3.0",
+    "react": "^17.0.2",
+    "react-dom": "^17.0.2",
     "react-redux": "^5.0.7",
-    "react-router": "^4.1.2",
-    "react-router-dom": "^4.1.2",
-    "react-router-redux": "^5.0.0-alpha.6",
-    "react-scripts": "5.0.1",
+    "react-router-dom": "^6.9.0",
+    "react-scripts": "^5.0.1",
     "redux": "^3.6.0",
     "redux-devtools-extension": "^2.13.2",
-    "redux-logger": "^3.0.1",
+    "sass": "^1.45.0",
     "superagent": "^3.8.2",
-    "superagent-promise": "^1.1.0"
+    "superagent-promise": "^1.1.0",
+    "typescript": "^4.8.2"
   },
   "scripts": {
     "start": "REACT_APP_WILCO_ID=${WILCO_ID:-\"$(cat ../.wilco)\"} react-scripts start",
@@ -60,5 +67,8 @@
       "last 1 firefox version",
       "last 1 safari version"
     ]
+  },
+  "resolutions": {
+    "autoprefixer": "10.4.5"
   }
 }
