@@ -2,12 +2,14 @@ import ItemActions from "./ItemActions";
 import { Link } from "react-router-dom";
 import React from "react";
 
-const ItemMeta = ({ item, canModify }) => {
+const ItemMeta = (props) => {
+  const item = props.item;
   return (
-    <div className="d-flex flex-row align-items-center pt-2">
+    <div id="item-meta" className="d-flex flex-row align-items-center pt-2">
       <Link to={`/@${item.seller.username}`}>
         <img
-          src={item.seller?.image}
+          id="seller-image"
+          src={item.seller.image}
           alt={item.seller.username}
           className="user-pic mr-2"
         />
@@ -20,7 +22,7 @@ const ItemMeta = ({ item, canModify }) => {
         <span className="date">{new Date(item.createdAt).toDateString()}</span>
       </div>
 
-      <ItemActions canModify={canModify} item={item} />
+      <ItemActions canModify={props.canModify} item={item} />
     </div>
   );
 };
