@@ -33,10 +33,11 @@ const ItemPreview = (props) => {
     <div
       className="card bg-dark border-light p-3"
       style={{ borderRadius: "20px" }}
+      id={`item_${item.slug}`}
     >
       <img
         alt="item"
-        src={item?.image || "placeholder.png"}
+        src={item.image}
         className="card-img-top item-img"
         style={{ borderRadius: "20px" }}
       />
@@ -45,10 +46,10 @@ const ItemPreview = (props) => {
           <h3 className="card-title">{item.title}</h3>
           <p className="card-text crop-text-3">{item.description}</p>
         </Link>
-        <div className="d-flex flex-row align-items-center pt-2">
+        <div className="d-flex flex-row align-items-center pt-2 item-footer">
           <Link to={`/@${item.seller.username}`} className="flex-grow-1">
             <img
-              src={item.seller?.image}
+              src={item.seller.image}
               alt={item.seller.username}
               className="user-pic rounded-circle pr-1"
             />
